Feature/refactor (#83)
* Refactored home page

* Created a constants file
