+++ lib/views/go_to_page.dart
@@ -187,7 +187,10 @@ class _GoToPageState extends State<GoToPage> {
                   backgroundColor: Color.fromRGBO(147, 0, 47, 1),
                   elevation: 2,
                   onPressed: () {
-                    if (originDestinationCoords[0] != null && originDestinationCoords[1] != null && originDestinationCoords[0] != originDestinationCoords[1]) {
+                    if (originDestinationCoords[0] != null &&
+                        originDestinationCoords[1] != null &&
+                        originDestinationCoords[0] !=
+                            originDestinationCoords[1]) {
                       Navigator.pop(context);
                       widget.coordinates(originDestinationCoords);
                     }
